Move data_format into own directory and rename
This relocates the command "derive-data-format" into it's own directory
to make it cleaner and easier to add new commands to the `snabb config`
program. This also renames it from "derive-data-format" to simply be
"data-format".
